renames package extractor => extractors
