+++ src/dimcat/__init__.py
@@ -13,7 +13,7 @@ finally:
 
 from .base import DimcatConfig, get_class, get_schema
 from .data.dataset.base import Dataset
-from .steps.extractor.base import FeatureExtractor
+from .steps.extractors.base import FeatureExtractor
 from .steps.loaders.base import PackageLoader
 from .steps.pipeline.base import Pipeline
