+++ src/dimcat/steps/extractors/__init__.py
@@ -2,4 +2,6 @@ from __future__ import annotations
 
 import logging
 
+from .base import FeatureExtractor
+
 logger = logging.getLogger(__name__)
