+++ src/dimcat/steps/extractors/base.py
