+++ tests/test_pipeline.py
@@ -1,7 +1,7 @@
 from dimcat.base import DimcatConfig
 from dimcat.data.resource.dc import PieceIndex
 from dimcat.steps.base import FeatureProcessingStep
-from dimcat.steps.extractor.base import FeatureExtractor
+from dimcat.steps.extractors.base import FeatureExtractor
 from dimcat.steps.grouper.base import CustomPieceGrouper
 from dimcat.steps.pipeline.base import Pipeline
